fix : 매개변수 타입 수정

+++ project/numpy/optim/optimizer.py
@@ -2,7 +2,7 @@ import numpy as np
 
 
 class _Optimizer:
-    def __init__(self, params: list, lr: int):
+    def __init__(self, params: list, lr: float):
         self.lr = lr
         self.params = params
